[change]コンポーネント構成の変更 コンポーネント構成を整理するため、各機能別niFeaturesを作成してコンポーネントを配置
